Migrate docs.js to component and import into index.md
