+++ docs/index.md
@@ -1,0 +1,9 @@
+---
+id: _index
+title: Docs Overview
+hide_title: true
+---
+
+import Docs from "../src/components/docsHome";
+
+<Docs />
