+++ src/components/styles.module.css
@@ -1,0 +1,398 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+/**
+ * CSS files with the .module.css suffix will be treated as CSS modules
+ * and scoped locally.
+ */
+
+:root {
+  --black: #000;
+  --black-light: #434343;
+  --bright-cyan-50: rgba(63, 196, 254, 0.2);
+  --bright-cyan: #3fc4fe;
+  --dark-blue-grey: #211b50;
+  --dark-slate-blue: #2d2665;
+  --dodger-blue-10: rgba(63, 187, 254, 0.1);
+  --dodger-blue: #3fbbfe;
+  --dusk-blue: #412a94;
+  --greenblue-10: rgba(31, 216, 152, 0.1);
+  --greenblue: #1fd898;
+  --greyish-brown: #474747;
+  --pale-grey: #e8e6f5;
+  --periwinkle-50: rgba(136, 89, 255, 0.2);
+  --periwinkle: #8759ff;
+  --sea-green-50: rgba(66, 255, 172, 0.2);
+  --sea-green: #42ffac;
+  --white: #fff;
+
+  --uni-background-color: var(--white);
+  --uni-heading-color: var(--dusk-blue);
+  --uni-heading-helper-color: var(--dusk-blue);
+  --uni-hero-bgcolor: var(--dark-blue-grey);
+  --uni-oval-background-color: var(--white);
+  --uni-oval-shadow-color: var(--pale-grey);
+  --uni-testimonial-text-color: var(--greyish-brown);
+}
+
+.heroBanner {
+  padding: 4rem 0;
+  text-align: center;
+  position: relative;
+  overflow: hidden;
+  min-height: 650px;
+  background-color: #807d7c !important;
+}
+
+.buttons {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.headerButtons {
+  border-radius: 5;
+  margin-bottom: 1rem;
+  color: #fff !important;
+}
+
+.quickstart {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 1em 0 0 1em;
+}
+
+.github {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 0 1em 1em 0;
+}
+
+.buttons-tools-github {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 0 1em 1em 0 !important;
+}
+
+.doc {
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+  display: flex;
+  border-bottom-left-radius: 100% 600px;
+  border-bottom-right-radius: 100% 600px;
+}
+
+.features {
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+  display: flex;
+  text-align: center;
+  justify-content: center;
+  border-bottom-left-radius: 100% 600px;
+  border-bottom-right-radius: 100% 600px;
+}
+
+.toolsback {
+  border-top-left-radius: 100% 100px;
+  border-top-right-radius: 100% 100px;
+  border-top: 5px solid var(--ifm-color-primary) !important;
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+  display: flex;
+  color: white;
+  text-align: center;
+  justify-content: center;
+  background-color: #bdc3c7;
+}
+
+.tools {
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+  display: flex;
+  color: white;
+  text-align: center;
+  justify-content: center;
+}
+
+.heroLogo {
+  animation-duration: 2s;
+  animation-name: jackInTheBox;
+  float: left;
+  height: 15rem;
+  margin-top: 1px;
+  padding: 5px;
+  padding-right: 5rem;
+}
+
+@media only screen and (max-width: 768px) {
+  .heroBanner {
+    padding-top: 2rem !important;
+  }
+
+  .heroInner {
+    padding: 0;
+  }
+
+  .heroProjectTagline {
+    font-size: 36px !important;
+    text-align: center !important;
+  }
+
+  .heroLogo {
+    display: block;
+    float: none;
+    margin: 0 auto;
+    height: 8rem;
+    padding: 5px;
+  }
+
+  .indexCtas {
+    text-align: center;
+  }
+
+  .indexCtasGitHubButton {
+    display: none;
+  }
+}
+
+.headerImage {
+  height: 300px;
+  width: 200px;
+  max-width: 100%;
+  max-height: 100%;
+  width: auto;
+  padding: 1rem 0;
+}
+
+.heroProjectTagline {
+  color: #fff;
+  font-size: 4rem;
+  margin: 0;
+  margin-top: 1rem;
+  text-align: left;
+  padding-bottom: 2rem;
+}
+
+.heroProjectKeywords {
+  color: var(--ifm-color-black);
+}
+
+.featureImage {
+  height: 150px;
+  width: 100px;
+  max-width: 100%;
+  max-height: 100%;
+  width: auto;
+  padding: 1rem 0;
+}
+
+.toolImage {
+  height: 80px;
+  max-width: 100%;
+  max-height: 100%;
+  width: auto;
+  padding: 1rem 0;
+}
+
+.particles {
+  background-color: "#262941";
+  z-index: -10;
+  position: "absolute";
+  width: "100%";
+  height: "100%";
+  top: 0;
+  left: 0;
+}
+.hero.hero--primary {
+  background-color: #807d7c !important;
+}
+
+.hero__title {
+  z-index: 1 !important;
+  color: #000000 !important;
+}
+
+.hero__subtitle {
+  z-index: 1 !important;
+  color: #a4a9ad !important;
+}
+
+.text__gray {
+  color: #000000 !important;
+}
+
+/* Animation */
+@keyframes jackInTheBox {
+  from {
+    opacity: 0;
+    transform: scale(0.1) rotate(30deg);
+    transform-origin: center bottom;
+  }
+
+  50% {
+    transform: rotate(-10deg);
+  }
+
+  70% {
+    transform: rotate(3deg);
+  }
+
+  to {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
+
+.text {
+  font-family: Poppins, sans-serif;
+  font-stretch: normal;
+  font-style: normal;
+  letter-spacing: normal;
+  line-height: normal;
+  color: var(--ifm-font-base-color);
+}
+
+.text-big {
+  font-size: 1.5em;
+  font-weight: bold;
+}
+
+.text-huge {
+  font-size: 2.5em;
+  font-weight: bold;
+  line-height: 1.25;
+}
+
+.text.heading-helper {
+  color: var(--uni-heading-helper-color);
+  font-family: Muli;
+  font-size: 1em;
+  font-weight: 600;
+  letter-spacing: 1px;
+  opacity: 0.6;
+}
+
+.heading {
+  color: var(--uni-heading-color);
+  font-family: Poppins;
+  font-size: 2em;
+  font-weight: bold;
+  line-height: 1.31;
+  margin-bottom: 2em;
+}
+
+.heading-white {
+  --uni-heading-color: var(--white);
+}
+
+.text.emphasis {
+  color: var(--dodger-blue);
+  font-family: Poppins;
+  font-size: 24px;
+  font-weight: 600;
+  line-height: 1.42;
+}
+
+.text.paragraph {
+  color: var(--white);
+  font-family: Muli;
+  font-size: 1em;
+  line-height: 1.75;
+  margin-bottom: 0.8em;
+  opacity: 0.8;
+}
+
+.text.section-subtitle {
+  color: var(--white);
+  font-family: Muli;
+  font-size: 1em;
+  font-weight: 600;
+  letter-spacing: 1px;
+  opacity: 0.6;
+}
+
+.text.footer-heading {
+  color: var(--dodger-blue);
+  font-family: Muli;
+  font-size: 18px;
+  font-weight: 600;
+}
+
+.supported {
+  color: var(--white);
+  font-family: Muli;
+  font-size: 0.85em;
+  font-weight: 600;
+  line-height: 2;
+  margin-top: 3em;
+  opacity: 0.6;
+}
+
+.title {
+  color: var(--white);
+  display: block;
+}
+
+.description {
+  color: var(--bright-cyan);
+  display: block;
+  margin: 24px 0;
+}
+
+.bullets {
+  color: var(--periwinkle);
+}
+
+.bullets li {
+  margin-bottom: 0;
+}
+
+.bullets li::before {
+  background-color: var(--periwinkle) !important;
+}
+
+@supports (background-clip: text) {
+  .description {
+    background-image: linear-gradient(
+      0.25turn,
+      var(--ifm-color-primary),
+      #807d7c
+    );
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+  }
+}
+
+.landing :global .hero .container {
+  margin-top: 0em;
+  margin-bottom: 0em;
+}
+
+.landing :global .container {
+  margin-top: 4em;
+  margin-bottom: 4em;
+}
+
+.landing .section-content :global .row {
+  position: relative;
+}
+
+@media (max-width: 996px) {
+  .landing :global .col {
+    margin-bottom: 2em;
+  }
+  .landing :global .col:last-of-type {
+    margin-bottom: 0;
+  }
+}
